+++ R/cNORM.R
@@ -48,8 +48,6 @@
 #'   \item CDC (2012). National Health and Nutrition Examination Survey: Questionnaires, Datasets
 #'   and Related Documentation. available: https://wwwn.cdc.gov/nchs/nhanes/OtherNhanesData.aspx.
 #'   date of retrieval: 25/08/2018
-#'   \item Harrel, F. (2020). Hmisc: Harrell Miscellaneous (v. 4.4-1). available https://CRAN.R-project.org/package=Hmisc
-#'   (code for weighted ranking adapted from wtd.rank & wtd.table by courtesy of Frank Harrell)
 #'   \item Lenhard, A., Lenhard, W., Suggate, S. & Segerer, R. (2016). A continuous solution to
 #'   the norming problem. Assessment, Online first, 1-14. doi: 10.1177/1073191116656437
 #'   \item Lenhard, A., Lenhard, W., Segerer, R. & Suggate, S. (2015). Peabody Picture Vocabulary
@@ -129,14 +127,14 @@ cNORM.GUI <- function(launch.browser=TRUE){
 #'
 #' Conducts continuous norming in one step and returns an object including ranked raw data and the continuous norming
 #' model. Please consult the function description ' of 'rankByGroup', 'rankBySlidingWindow' and 'bestModel' for specifics
-#' of the steps in the data preparation and modelling process. In addition to the raw scores, either provide
+#' of the steps in the data preparation and modeling process. In addition to the raw scores, either provide
 #' \itemize{
 #'  \item{a numeric vector for the grouping information (group)}
 #'  \item{a numeric vector for both grouping and age (group, age)}
 #'  \item{a numeric age vector and the width of the sliding window (age, width)}
 #' }
 #' for the ranking of the raw scores. You can
-#' adjust the grade of smoothing of the regression modell by setting the k and terms parameter. In general,
+#' adjust the grade of smoothing of the regression model by setting the k and terms parameter. In general,
 #' increasing k to more than 4 and the number of terms lead to a higher fit, while lower values lead to more
 #' smoothing.
 #' @param raw Numeric vector of raw scores
@@ -154,10 +152,11 @@ cNORM.GUI <- function(launch.browser=TRUE){
 #' ranking order with higher raw scores getting lower norm scores; relevant
 #' for example when norming error scores, where lower scores mean higher
 #' performance
-#' @param weights Vector or variable name in the dataset with weights to compensate imbalances due to insufficient norm
-#' data stratification. All weights have to be numerical and positive. The code to compute weighted percentiles originates from the
-#' Hmisc package (functions) wtd.rank and wtd.table) and is provided by the courtesy of Frank Harrell. Please note, that this
-#' feature is currently EXPERIMENTAL!
+#' @param weights Vector or variable name in the dataset with weights for each individual case. It can be used
+#' to compensate for moderate imbalances due to insufficient norm data stratification. Weights should be numerical
+#' and positive.
+#' Please note, that this feature is currently EXPERIMENTAL and subject to ongoing work! Precision of weighting increases
+#' with sample size. On the other hand, in large samples, it is easy to stratificate and then weighting is not needed anymore.
 #' @param terms Selection criterion for model building. The best fitting model with
 #' this number of terms is used
 #' @param R2 Adjusted R square as a stopping criterion for the model building
@@ -180,12 +179,20 @@ cNORM.GUI <- function(launch.browser=TRUE){
 #' # ... or instead of raw scores for norm scores, the other way round
 #' rawTable(c(2.125, 2.375, 2.625, 2.875), cnorm.elfe, CI = .90, reliability = .95)
 #'
+#'
+#' # Not really a plausible scenario, but just for demonstration purposes, we will
+#' # use the PPVT dataset and sex as the weighting variable (1 = male, 2 = female),
+#' # and consequently, females will get the double weight. This procedure can be used
+#' # to correct imbalances in the dataset, but it is still experimental. Please use
+#' # positive, non-zero numerics, preferably integers for this:
+#' \dontrun{
+#' cnorm.ppvt <- cnorm(raw = ppvt$raw, group = ppvt$group, weight = ppvt$sex)
+#' }
+#'
 #' @export
 #' @references
 #' \enumerate{
 #'   \item Gary, S. & Lenhard, W. (2021). In norming we trust. Diagnostica.
-#'   \item Harrel, F. (2020). Hmisc: Harrell Miscellaneous (v. 4.4-1). available https://CRAN.R-project.org/package=Hmisc
-#'   (code for weighted ranking adapted from wtd.rank & wtd.table by courtesy of Frank Harrell)
 #'   \item Lenhard, A., Lenhard, W., Suggate, S. & Segerer, R. (2016). A continuous solution to the norming problem. Assessment, Online first, 1-14. doi:10.1177/1073191116656437
 #'   \item Lenhard, A., Lenhard, W., Gary, S. (2018). Continuous Norming (cNORM). The Comprehensive R Network, Package cNORM, available: https://CRAN.R-project.org/package=cNORM
 #'   \item Lenhard, A., Lenhard, W., Gary, S. (2019). Continuous norming of psychometric tests: A simulation study of parametric and semi-parametric approaches. PLoS ONE, 14(9),  e0222279. doi:10.1371/journal.pone.0222279
@@ -204,6 +211,10 @@ cnorm <- function(raw = NULL,
                   k = 4,
                   terms = 0,
                   R2 = NULL){
+
+  if(!is.null(weights)){
+    warning("Weighting is still an experimental feature. It is currently not yet encouraged to use it.")
+  }
 
   if(is.numeric(raw)&&is.numeric(group)){
     if(length(raw)!=length(group)){