+++ NEWS.md
@@ -3,7 +3,23 @@ This file documents the development of the package as well as open issues or poi
 
 
 
-### Version in 2.0.0 (release candidate)
+### Version in 2.0.1 (release candidate)
+Date: 2021.01.05
+
+
+Changes:
+
+*    Fixing errors in the context of weighted percentile modelling
+*    Code change of weighted rank estimation from https://aakinshin.net/posts/weighted-quantiles/
+     code by Andrey Akinshin
+*    Additional message for plotting, when weighted percentiles are used
+*    Use weighted percentiles in plotPercentiles
+*    automatic weighting deactivated in bestModel, since it is already applied in ranking
+*    suppressWarnings in weighted ranking
+
+
+
+### Version in 2.0.0 (release)
 Date: 2020.12.04
 Version 2.0.0 features many fundamental improvements both relating to the procedure but as well to the
 package itself. It introduces weighted percentiles and thus helps in correcting violations of